moved hexpansion EEPROM code to EEPROM sub-directory
