+++ tests/conftest.py
@@ -162,7 +162,7 @@ def install_fake_hexpansion(vid: int, pid: int, port: int,
     if app_class is None:
         app_class = HexDriveApp
     if app_version is None:
-        from sim.apps.BadgeBot.hexdrive import VERSION
+        from sim.apps.BadgeBot.EEPROM.hexdrive import VERSION
         app_version = VERSION
 
     fake_app = app_class(port, app_version)
